requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
databricks-cli==0.18.0
idna==3.10
oauthlib==3.3.1
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
tabulate==0.9.0
urllib3==2.5.0
